% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_name_new.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regular expression special characters}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{A character vector of strings to escape.}
}
\value{
A character vector with regex metacharacters escaped.
}
\description{
Escapes all regex metacharacters in a character vector so that the resulting
strings can be safely used in regular expression patterns.
}
\details{
The following characters are escaped: \code{[]{}()+*^$|\\?.}
}
\examples{
escape_regex("a+b")   # returns "a\\+b"
escape_regex("file?.txt")  # returns "file\\?\\.txt"

}
